Updated implementation.
